+++ govtool/backend/app/Main.hs
@@ -8,7 +8,6 @@
 
 module Main where
 
-import           Control.Concurrent.QSem                (newQSem)
 import           Control.Exception                      (Exception, SomeException, fromException, throw)
 import           Control.Lens.Operators                 ((.~))
 import           Control.Monad
@@ -38,8 +37,6 @@ import qualified Data.Text.Lazy.Encoding                as LazyText
 
 import           Database.PostgreSQL.Simple             (close, connectPostgreSQL)
 
-import           Network.HTTP.Client                    hiding (Proxy, Request)
-import           Network.HTTP.Client.TLS
 import           Network.Wai
 import           Network.Wai.Handler.Warp
 import           Network.Wai.Middleware.Cors
@@ -85,6 +82,7 @@ startApp vvaConfig = do
       settings =
         setPort vvaPort
           $ setHost vvaHost
+          $ setTimeout 60 -- 60 seconds timeout
           $ setBeforeMainLoop
             ( Text.hPutStrLn stderr $
                 Text.pack
@@ -107,8 +105,6 @@ startApp vvaConfig = do
     dRepVotingPowerCache <- newCache
     dRepListCache <- newCache
     networkMetricsCache <- newCache
-    proposalMetadataValidationCache <- newCache
-    dRepMetadataValidationCache <- newCache
     return $ CacheEnv
       { proposalListCache
       , getProposalCache
@@ -120,13 +116,9 @@ startApp vvaConfig = do
       , dRepVotingPowerCache
       , dRepListCache
       , networkMetricsCache
-      , proposalMetadataValidationCache
-      , dRepMetadataValidationCache
       }
   connectionPool <- createPool (connectPostgreSQL (encodeUtf8 (dbSyncConnectionString $ getter vvaConfig))) close 1 1 60
-  vvaTlsManager <- newManager tlsManagerSettings
-  qsem <- newQSem (metadataValidationMaxConcurrentRequests vvaConfig)
-  let appEnv = AppEnv {vvaConfig=vvaConfig, vvaCache=cacheEnv, vvaConnectionPool=connectionPool, vvaTlsManager, vvaMetadataQSem=qsem}
+  let appEnv = AppEnv {vvaConfig=vvaConfig, vvaCache=cacheEnv, vvaConnectionPool=connectionPool }
   server' <- mkVVAServer appEnv
   runSettings settings server'