+++ govtool/frontend/src/hooks/queries/useGetDRepVotesQuery.ts
@@ -33,7 +33,7 @@ export const useGetDRepVotesQuery = (
   });
 
   const groupedByType = data?.reduce((groups, item) => {
-    const itemType = item.proposal.type;
+    const itemType = item?.proposal.type;
 
     // TODO: Provide better typing for groups
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment